Replace Travis with GitHub Actions.

+++ scripts/publish.sh
@@ -1,22 +1,25 @@
 #!/bin/bash
-set -u
+set -eou pipefail
 
 GENERATOR="${1}"
 
-DIRECTORY="public"
-BRANCH="gh-pages"
+GITHUB_TOKEN="${2}"
+GITHUB_REPO="${3}"
+GITHUB_BRANCH="gh-pages"
 
-echo ":: Removing [${DIRECTORY}]."
+DIRECTORY="public"
+
+echo ":: Removing [${DIRECTORY}] directory."
 rm -rf "${DIRECTORY}"
 
-echo ":: Removing [${DIRECTORY}] working tree."
+echo ":: Removing [${DIRECTORY}] Git working tree."
 git worktree remove "${DIRECTORY}"
 
-echo ":: Re-creating [${DIRECTORY}] working tree."
-git worktree add -B "${BRANCH}" "${DIRECTORY}" origin/"${BRANCH}"
+echo ":: Re-creating [${DIRECTORY}] Git working tree."
+git worktree add -B "${GITHUB_BRANCH}" "${DIRECTORY}" origin/"${GITHUB_BRANCH}"
 
-echo ":: Removing [${DIRECTORY}] working tree contents."
-rm -rf "${DIRECTORY}"/*
+echo ":: Removing [${DIRECTORY}] Git working tree contents."
+rm -rf "${DIRECTORY:?}"/*
 
 eval "${GENERATOR}"
 
@@ -24,7 +27,7 @@ echo ":: Committing generated contents."
 cd "${DIRECTORY}"
 git add --all
 
-if [[ -n "${TRAVIS}" ]]; then
+if [[ -n "${CI}" ]]; then
   git config user.name "Publisher"
   git config user.email "publisher@localhost"
 fi
@@ -32,9 +35,9 @@ fi
 git commit --message "Publish Hugo-generated contents."
 
 echo ":: Pushing..."
-if [[ -z "${TRAVIS_REPO_SLUG}" || -z "${GITHUB_TOKEN}" ]]; then
-  git push origin "${BRANCH}"
+if [[ -z "${GITHUB_TOKEN}" || -z "${GITHUB_REPO}" ]]; then
+  git push origin "${GITHUB_BRANCH}"
 else
-  git push --quiet "https://${GITHUB_TOKEN}:x-oauth-basic@github.com/${TRAVIS_REPO_SLUG}.git" "${BRANCH}"
+  git push --quiet "https://${GITHUB_TOKEN}:x-oauth-basic@github.com/${GITHUB_REPO}.git" "${GITHUB_BRANCH}"
 fi
